python_backend/requirements.txt
fastapi
uvicorn
openai
pydantic
python-multipart
playwright
Pillow
PyMuPDF
easyocr
google-generativeai
python-dotenv
